Move compositeSchema abstraction up to componsiteStore.
Promote the composite schema abstraction to "composite chunk store" - a chunk store which delegates to different chunk stores based on time.  This allows us to vary the store implementation over time, and not just the schema.  This will unblock the new bigtable storage adapter (using columns instead of rows), and allow us to more easily implement the iterative intersections and indexing of series instead of chunks.

Corner case when a writing chunks which span multiple stores: they are written to both stores, and instead of using the chunk start/end we use the schema start/end.  This will lead to duplication of the index entries on schema migrations, but is actually already the case for day boundaries anyway.  It will lead to duplicate writes of the chunk on schema migrations - they should be deduped by the underlying store.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>